refactor: Reworked the OutboundPackets class
This class needed to have a bit of a shuffle around to add in support
for Xenon gas. So I took the opportunity to rearrange the way the
packets and their IDs are defined, to make adding more of these
much quicker down the line.
Now, not all of the IDs will need to be reassigned, every time that
a new OutboundPacket is added.