docs: Updating BUILDING.md with changes to symlinking
Ended up changing over to using Windows Symlinks, so that one type of
symlink could be used for both VS Code and WSL. This is because Windows
can use it's own symlinks but not WSL's ones, and WSL can use either
Windows's symlinks, or its own.
I did keep the Linux instructions in, to make it easier for  those
building on Linux.