refactor: Improved the way the terminal commands are loaded
Previously, the way the terminal commands worked meant that they took
presidence over the main KerbalSimpit code. There was also a lot of
rather ugly OOP code written by a rather less experianced me. These
tweaks to the terminal commands for the mod will make the future
maintenance of the codebase much simpler, as well as making it
clearer to onlookers what is going on.