fix: Fixed reconnecting controllers not binding
These changes should have fixed an issue that we were facing in regards
to reconnecting a controller while in the flight scene. It appears that
the event dispatch thread was not being restarted when the serial start
command was being issued by a user in the terminal.

Also changed the serial start command to actually call the intended
method when the command is issued.