refactor: Various changes to compact/improve the flow, and to fix potential bugs
Changed the WarpControlValues to match their refactored names.
Changed the switch statement for all of the WarpControlValues to
use multiple cases per switch statement where possible.

Changed the way the warp rates are set, to prevent an issue where
if another mod changes the timewarp values, our mod would not
be able to set timewarp values.