feat: Upgraded the default cfg file generation
Previously the default cfg file generation was not aware of the OS that
it was running on, and had a hardcoded value for a Unix serial port.
This change makes it so that the mod attempts to pick the right
default serial port value to be used, based on the OS the game is
being played on. This is yet to be tested on a Unix based system, to
make sure that detection is working correctly, however it does work for
Windows - which is what the majority of users will be running it on
anyways.