feat: Created a Windows install batch script
Created a basic script to copy the mod files over to the specified
game install directory, to be used on windows. Requires the symlinks
to be setup, to be able to correctly do this.

Now this just needs to be tied into the csproj file, so that
it can be executed by Visual Studio.